==> 3_Contextual_features.ipynb <==
import sys
import csv
import time
import itertools
import numpy as np
import pandas as pd
from tqdm import tqdm
from tqdm import tqdm_notebook
from sklearn.neighbors import NearestNeighbors
from scipy import spatial
import datetime
from sklearn.linear_model import LogisticRegression
from ast import literal_eval
from sklearn import preprocessing

def contextual(pt,st):
    df=pd.read_csv( str(pt) + '/final_sorted_50.csv' , header=0,sep=',' )
    df=df.set_index('user_id')
    #remove zero rows 
    df=df[(df.f1+df.f2+df.f3+df.f4+df.f5+df.f6+df.f7+df.f8 +df.f9 +df.f10 +df.f11+df.f12+df.f13 +df.f14 +df.f15 != 0)]
    #preparation of contextual features for users
    df_us=pd.read_csv('/Users/amirasarbaev/Downloads/Internship/data/users.csv',header=0,sep='\t')
    df_us['career_level'].fillna(0, inplace=True)
    df_us['region'].fillna(0, inplace=True)
    df_us['experience_n_entries_class'].fillna(0, inplace=True)
    df_us['experience_years_experience'].fillna(0, inplace=True)
    df_us['experience_years_in_current'].fillna(0, inplace=True)
    df_us['edu_degree'].fillna(0, inplace=True)
    df_us['edu_fieldofstudies'].fillna(0, inplace=True)
    df_us=df_us.set_index('id')
    df_us.index.names = ['user_id']
    df_us["country"] = df_us["country"].astype('category')
    df_us["country"] = df_us["country"].cat.codes
    df_us=df_us[['career_level','discipline_id','industry_id','country','region','experience_n_entries_class','experience_years_experience','experience_years_in_current','edu_degree']]
    #we are going to merget statistical features with users contextual
    df=pd.merge(df, df_us, left_index=True, right_index=True)
    #preparation of contextual features for items
    df_it=pd.read_csv('/Users/amirasarbaev/Downloads/Internship/data/items.csv',header=0,sep='\t')
    df_it.fillna(0, inplace=True)
    df_it=df_it.set_index('id')
    df_it.index.names = ['item_id'] 
    df_it["country"] = df_it["country"].astype('category')
    df_it["country_cat"]= df_it["country"].cat.codes
    df_it=df_it[['career_level','discipline_id','industry_id','country_cat','region','employment','active_during_test']]
    #we are going to merget statistical features with offers contextual
    df=pd.merge(df, df_it, left_on='item_id', right_index=True)
    df.to_csv(str(st) + "/3_final_dataset.csv")
    return df

contextual("/Users/amirasarbaev/Desktop","/Users/amirasarbaev/Desktop")
