==> 4_Time_line.ipynb <==
import sys
import csv
import time
import itertools
import numpy as np
import pandas as pd
from tqdm import tqdm_notebook
from sklearn import datasets
import matplotlib.pyplot as plt 
from sklearn.metrics import f1_score
from sklearn.feature_selection import RFE
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LogisticRegressionCV
from sklearn.model_selection import train_test_split
from sklearn import model_selection
from sklearn.model_selection import cross_val_score
from sklearn.metrics import classification_report

from sklearn import preprocessing
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import Normalizer

def time_line(pt,st):
    df=pd.read_csv( str(pt) + '/3_final_dataset.csv',header=0,sep=',')
    df_int=pd.read_csv( str(pt) + '/1_target_list.csv',header=0,sep=',')
    #remove duplicates to keep a time of earliest interaction.
    df_click=df_int[df_int.interaction_type==1]
    df_click.drop_duplicates(keep='first')
    df_book=df_int[df_int.interaction_type==2]
    df_book.drop_duplicates(keep='first')
    df_reply=df_int[df_int.interaction_type==3]
    df_reply.drop_duplicates(keep='first')
    df_undisp=df_int[df_int.interaction_type==0]
    
    df_click=df_click.groupby(['user_id','item_id'])['created_at'].apply(lambda x: np.min(x)).reset_index()
    df_book=df_book.groupby(['user_id','item_id'])['created_at'].apply(lambda x: np.min(x)).reset_index()
    df_reply=df_reply.groupby(['user_id','item_id'])['created_at'].apply(lambda x: np.min(x)).reset_index()
    df_undisp=df_undisp.groupby(['user_id','item_id'])['created_at'].apply(lambda x: np.min(x)).reset_index()
    
    df=pd.merge(df,df_click,how='left', on=['user_id','item_id'])
    df=df.rename(columns = {'created_at':'week_click'})
    df=pd.merge(df,df_book,how='left', on=['user_id','item_id'])
    df=df.rename(columns = {'created_at':'week_book'})
    df=pd.merge(df,df_reply,how='left', on=['user_id','item_id'])
    df=df.rename(columns = {'created_at':'week_reply'})
    df=pd.merge(df,df_undisp,how='left', on=['user_id','item_id'])
    df=df.rename(columns = {'created_at':'week_undisp'})
    
    df.week_click.fillna(df.week_undisp, inplace=True)
    df.week_book.fillna(df.week_undisp, inplace=True)
    df.week_reply.fillna(df.week_undisp, inplace=True)
    
    df.set_index(['user_id', 'item_id'], inplace=True)
    
    df.to_csv(str(st)+'/4_time_line.csv')
    return df

time_line("/Users/amirasarbaev/Desktop","/Users/amirasarbaev/Desktop")
